requirements.txt
streamlit
pillow
ipyleaflet==0.16.0
bokeh==2.4.3